Moved test_vectors.h to tests/crypto. Matching the signature validity with expected results.
