+++ tests/CMakeLists.txt
@@ -14,7 +14,7 @@ function(add_iota_test test_src test_name)
     target_link_libraries(${test_name} PRIVATE iota_core unity)
   elseif(${test_src} MATCHES "^crypto")
     add_dependencies(${test_name} iota_crypto)
-    target_link_libraries(${test_name} PRIVATE iota_crypto unity)
+    target_link_libraries(${test_name} PRIVATE iota_crypto iota_core unity)
   else()
     add_dependencies(${test_name} iota_client)
     target_link_libraries(${test_name} PRIVATE iota_client unity)
